+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the Multisite Exporter plugin will be documented in this 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.3] - 2025-05-14
+### Fixed
+- WP-CLI package installation and command execution issues
+- Improved compatibility for `wp package install` by using anonymous functions
+
 ## [1.2.2] - 2025-05-14
 ### Added
 - Support for installing WP-CLI commands as a package with `wp package install`